+++ BovineLabs.Core.Extensions.Authoring/ObjectManagement/ObjectManagementSettingsBase.cs
@@ -10,22 +10,19 @@ namespace BovineLabs.Core.Authoring.ObjectManagement
     using System.Linq;
     using BovineLabs.Core.Assertions;
     using BovineLabs.Core.Authoring.Settings;
-    using BovineLabs.Core.Extensions;
     using BovineLabs.Core.ObjectManagement;
     using Unity.Entities;
     using UnityEngine;
 
     public abstract class ObjectManagementSettingsBase : SettingsBase
     {
-        private Dictionary<GameObject, int>? objectDefinitionMap;
-
-        public abstract int Mod { get; }
+        private Dictionary<GameObject, uint>? objectDefinitionMap;
 
         public abstract IReadOnlyCollection<ObjectDefinition> ObjectDefinitions { get; }
 
         public abstract IReadOnlyCollection<ObjectGroup> ObjectGroups { get; }
 
-        public Dictionary<GameObject, int> ObjectDefinitionMap
+        public Dictionary<GameObject, uint> ObjectDefinitionMap
         {
             get
             {
@@ -33,14 +30,14 @@ namespace BovineLabs.Core.Authoring.ObjectManagement
 
                 if (this.objectDefinitionMap == null)
                 {
-                    var notNull = this.ObjectDefinitions.Where(o => o.Prefab != null).ToArray();
+                    var notNull = this.ObjectDefinitions.Where(o => o.Prefab).ToArray();
                     var distinct = notNull.Distinct(default(PrefabDistinct)).ToArray();
                     if (distinct.Length != notNull.Length)
                     {
                         BLGlobalLogger.LogErrorString("Non-unique object definitions. Make a prefab instance if you need to duplicate one");
                     }
 
-                    this.objectDefinitionMap = distinct.ToDictionary(o => o.Prefab!, o => o.ID);
+                    this.objectDefinitionMap = distinct.ToDictionary(o => o.Prefab!, o => (uint)o.ID);
                 }
 
                 return this.objectDefinitionMap;
@@ -85,20 +82,15 @@ namespace BovineLabs.Core.Authoring.ObjectManagement
                 return;
             }
 
-            baker.AddComponent(entity, new Mod(this.Mod));
             var registry = baker.AddBuffer<ObjectDefinitionSetupRegistry>(entity);
 
             var definitions = this.ObjectDefinitions.Where(o => o).ToList();
-
-            // Get the largest ID and resize the array to fit
-            // The importer should keep this close to the number of prefabs
-            definitions.Sort((d1, d2) => d1.ID.CompareTo(d2.ID));
-            var length = definitions[^1].ID + 1;
-            registry.ResizeInitialized(length);
+            registry.Capacity = definitions.Count;
 
             foreach (var asset in definitions)
             {
                 baker.DependsOn(asset);
+
                 if (asset.ID != 0 && !asset.Prefab)
                 {
                     BLGlobalLogger.LogWarningString($"Missing Prefab on {asset}");
@@ -106,7 +98,11 @@ namespace BovineLabs.Core.Authoring.ObjectManagement
                 }
 
                 var prefab = baker.GetEntity(asset.Prefab, TransformUsageFlags.None);
-                registry[asset.ID] = new ObjectDefinitionSetupRegistry { Prefab = prefab };
+                registry.Add(new ObjectDefinitionSetupRegistry
+                {
+                    Id = new ObjectId(asset.ID),
+                    Prefab = prefab,
+                });
             }
         }
 
@@ -138,7 +134,7 @@ namespace BovineLabs.Core.Authoring.ObjectManagement
 
                 foreach (var definition in definitions)
                 {
-                    var id = new ObjectId(this.Mod, definition.ID);
+                    var id = new ObjectId(definition.ID);
 
                     objectGroupMatcher.Add((group, id));
                     objectGroupRegistry.Add(group, id);