+++ BovineLabs.Core.Editor/Dependency/ComponentDependencyWindow.cs
@@ -21,7 +21,7 @@ namespace BovineLabs.Core.Editor.Dependency
 
         protected override string DefaultButtonText => "Components";
 
-        [MenuItem("BovineLabs/Tools/Component Dependencies")]
+        [MenuItem(EditorMenus.RootMenuTools + "Component Dependencies")]
         public static void OpenWindow()
         {
             GetWindow<ComponentDependencyWindow>().Show();